requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
werkzeug==3.0.1
jinja2==3.1.2
sqlalchemy==2.0.23
click==8.1.7
itsdangerous==2.1.2
markupsafe==2.1.3 